Coe refactoring. Selected execption highlighting tests.
