+++ src/restrictededitingui.js
@@ -20,6 +20,13 @@ export default class RestrictedEditingUI extends Plugin {
 	/**
 	 * @inheritDoc
 	 */
+	static get pluginName() {
+		return 'RestrictedEditingUI';
+	}
+
+	/**
+	 * @inheritDoc
+	 */
 	init() {
 		const editor = this.editor;
 		const t = editor.t;
