+++ tests/restrictedediting.js
@@ -9,172 +9,36 @@ import testUtils from '@ckeditor/ckeditor5-core/tests/_utils/utils';
 import ClassicTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/classictesteditor';
 
 import RestrictedEditing from './../src/restrictedediting';
-import { setData as setModelData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model';
-import { getData as getViewData } from '@ckeditor/ckeditor5-engine/src/dev-utils/view';
-import VirtualTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/virtualtesteditor';
-import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph';
+import RestrictedEditingUI from './../src/restrictededitingui';
+import RestrictedEditingEditing from './../src/restrictededitingediting';
 
 describe( 'RestrictedEditing', () => {
 	let editor, element;
 
 	testUtils.createSinonSandbox();
 
-	describe( 'plugin', () => {
-		beforeEach( async () => {
-			element = document.createElement( 'div' );
-			document.body.appendChild( element );
+	beforeEach( async () => {
+		element = document.createElement( 'div' );
+		document.body.appendChild( element );
 
-			editor = await ClassicTestEditor.create( element, { plugins: [ RestrictedEditing ] } );
-		} );
-
-		afterEach( () => {
-			element.remove();
-
-			return editor.destroy();
-		} );
-
-		it( 'should be named', () => {
-			expect( RestrictedEditing.pluginName ).to.equal( 'RestrictedEditing' );
-		} );
-
-		it( 'should be loaded', () => {
-			expect( editor.plugins.get( RestrictedEditing ) ).to.be.instanceOf( RestrictedEditing );
-		} );
+		editor = await ClassicTestEditor.create( element, { plugins: [ RestrictedEditing ] } );
 	} );
 
-	describe( 'conversion', () => {
-		let model;
+	afterEach( () => {
+		element.remove();
 
-		beforeEach( async () => {
-			editor = await VirtualTestEditor.create( { plugins: [ Paragraph, RestrictedEditing ] } );
-			model = editor.model;
-		} );
-
-		afterEach( () => {
-			return editor.destroy();
-		} );
-
-		describe( 'upcast', () => {
-			it( 'should convert <span class="ck-restricted-editing-exception"> to marker', () => {
-				editor.setData( '<p>foo <span class="ck-restricted-editing-exception">bar</span> baz</p>' );
-
-				expect( model.markers.has( 'restricted-editing-exception:1' ) ).to.be.true;
-
-				const marker = model.markers.get( 'restricted-editing-exception:1' );
-
-				expect( marker.getStart().path ).to.deep.equal( [ 0, 4 ] );
-				expect( marker.getEnd().path ).to.deep.equal( [ 0, 7 ] );
-			} );
-
-			it( 'should convert multiple <span class="ck-restricted-editing-exception">', () => {
-				editor.setData(
-					'<p>foo <span class="ck-restricted-editing-exception">bar</span> baz</p>' +
-					'<p>ABCDEF<span class="ck-restricted-editing-exception">GHIJK</span>LMNOPQRST</p>'
-				);
-
-				expect( model.markers.has( 'restricted-editing-exception:1' ) ).to.be.true;
-				expect( model.markers.has( 'restricted-editing-exception:2' ) ).to.be.true;
-
-				// Data for the first marker is the same as in previous tests so no need to test it again.
-				const secondMarker = model.markers.get( 'restricted-editing-exception:2' );
-
-				expect( secondMarker.getStart().path ).to.deep.equal( [ 1, 6 ] );
-				expect( secondMarker.getEnd().path ).to.deep.equal( [ 1, 11 ] );
-			} );
-
-			it( 'should not convert other <span> elements', () => {
-				editor.setData( '<p>foo <span class="foo bar">bar</span> baz</p>' );
-
-				expect( model.markers.has( 'restricted-editing-exception:1' ) ).to.be.false;
-			} );
-		} );
-
-		describe( 'downcast', () => {
-			it( 'should convert model marker to <span>', () => {
-				setModelData( model, '<paragraph>foo bar baz</paragraph>' );
-
-				const paragraph = model.document.getRoot().getChild( 0 );
-
-				model.change( writer => {
-					writer.addMarker( 'restricted-editing-exception:1', {
-						range: writer.createRange( writer.createPositionAt( paragraph, 4 ), writer.createPositionAt( paragraph, 7 ) ),
-						usingOperation: true,
-						affectsData: true
-					} );
-				} );
-
-				const expectedView = '<p>foo <span class="ck-restricted-editing-exception">bar</span> baz</p>';
-				expect( editor.getData() ).to.equal( expectedView );
-				expect( getViewData( editor.editing.view, { withoutSelection: true } ) ).to.equal( expectedView );
-			} );
-
-			it( 'converted <span> should be the outermost attribute element', () => {
-				editor.conversion.for( 'downcast' ).attributeToElement( { model: 'bold', view: 'b' } );
-				setModelData( model, '<paragraph><$text bold="true">foo bar baz</$text></paragraph>' );
-
-				const paragraph = model.document.getRoot().getChild( 0 );
-
-				model.change( writer => {
-					writer.addMarker( 'restricted-editing-exception:1', {
-						range: writer.createRange( writer.createPositionAt( paragraph, 0 ), writer.createPositionAt( paragraph, 'end' ) ),
-						usingOperation: true,
-						affectsData: true
-					} );
-				} );
-
-				const expectedView = '<p><span class="ck-restricted-editing-exception"><b>foo bar baz</b></span></p>';
-				expect( editor.getData() ).to.equal( expectedView );
-				expect( getViewData( editor.editing.view, { withoutSelection: true } ) ).to.equal( expectedView );
-			} );
-		} );
+		return editor.destroy();
 	} );
 
-	describe( 'editing behavior', () => {
-		let model;
+	it( 'should be named', () => {
+		expect( RestrictedEditing.pluginName ).to.equal( 'RestrictedEditing' );
+	} );
 
-		beforeEach( async () => {
-			editor = await VirtualTestEditor.create( { plugins: [ Paragraph, RestrictedEditing ] } );
-			model = editor.model;
-		} );
+	it( 'should load the RestrictedEditingEditing plugin', () => {
+		expect( editor.plugins.get( RestrictedEditingEditing ) ).to.be.instanceOf( RestrictedEditingEditing );
+	} );
 
-		afterEach( () => {
-			return editor.destroy();
-		} );
-
-		it( 'should keep markers in the view when editable region is edited', () => {
-			setModelData( model,
-				'<paragraph>foo bar baz</paragraph>' +
-				'<paragraph>xxx y[]yy zzz</paragraph>'
-			);
-
-			const firstParagraph = model.document.getRoot().getChild( 0 );
-			const secondParagraph = model.document.getRoot().getChild( 1 );
-
-			model.change( writer => {
-				writer.addMarker( 'restricted-editing-exception:1', {
-					range: writer.createRange( writer.createPositionAt( firstParagraph, 4 ), writer.createPositionAt( firstParagraph, 7 ) ),
-					usingOperation: true,
-					affectsData: true
-				} );
-				writer.addMarker( 'restricted-editing-exception:2', {
-					range: writer.createRange(
-						writer.createPositionAt( secondParagraph, 4 ),
-						writer.createPositionAt( secondParagraph, 7 )
-					),
-					usingOperation: true,
-					affectsData: true
-				} );
-			} );
-
-			model.change( writer => {
-				model.insertContent( writer.createText( 'R', model.document.selection.getAttributes() ) );
-			} );
-
-			const expectedView = '<p>foo <span class="ck-restricted-editing-exception">bar</span> baz</p>' +
-				'<p>xxx <span class="ck-restricted-editing-exception">yRyy</span> zzz</p>';
-
-			expect( editor.getData() ).to.equal( expectedView );
-			expect( getViewData( editor.editing.view, { withoutSelection: true } ) ).to.equal( expectedView );
-		} );
+	it( 'should load the RestrictedEditingUI plugin', () => {
+		expect( editor.plugins.get( RestrictedEditingUI ) ).to.be.instanceOf( RestrictedEditingUI );
 	} );
 } );
